requirements.txt
flask
line-bot-sdk==2.4.1
aiohttp==3.8.3
